backend/requirements-base.txt
# Core dependencies (lightweight, rarely change)
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib==1.7.4
bcrypt==3.2.0
redis
python-multipart
httpx
prometheus-client
python-json-logger
psutil
